Set sending profile at template level

+++ src/api/views/sendingprofile_views.py
@@ -14,12 +14,13 @@ from api.serializers.sendingprofile_serializers import (
     SendingProfileDeleteResponseSerializer,
     SendingProfileSerializer,
 )
-from api.services import CampaignService
+from api.services import CampaignService, TemplateService
 from api.utils.subscription.campaigns import get_campaign_smtp
 
 # GoPhish API Manager
 campaign_manager = CampaignManager()
 campaign_service = CampaignService()
+template_service = TemplateService()
 
 
 class SendingProfilesListView(APIView):
@@ -100,6 +101,11 @@ class SendingProfileView(APIView):
 
     def delete(self, request, id):
         """Delete."""
+        if template_service.exists(parameters={"sending_profile_id": id}):
+            return Response(
+                {"error": "Templates are currently utilizing this sending profile."},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
         delete_response = campaign_manager.delete_sending_profile(smtp_id=id)
         serializer = SendingProfileDeleteResponseSerializer(delete_response)
         return Response(serializer.data, status=status.HTTP_200_OK)
